Forestfire Problem in Assembly
Mimicked the forestfire problem. This program takes in dimensions of a rectangle, number of generations, wind direction, and a grid containing seeds (.), burning trees (B), and healthy trees (t). Seeds spread in the direction of the wind and grow healthy trees each generation while burning trees infect any healthy trees that are directly next to them.